requirements.txt
fastapi
uvicorn
tortoise-orm
pydantic
mysqlclient
aiomysql
aerich
pyjwt
passlib
redis
jwt
datetime